[BTA] Ignore USE_FIR_RUNNER in the implementation
Second step of removing the FIR based incremental runner: nothing in the
Build Tools API implementation reads the option any more, so the JVM
incremental compiler is always the classic one. The option itself is
still declared in the public API and is removed in a later commit, which
keeps the API dump change reviewable on its own.

^KT-86873

# Conflicts:
#	compiler/build-tools/kotlin-build-tools-impl/src/main/kotlin/org/jetbrains/kotlin/buildtools/internal/CompilationServiceImpl.kt
#	compiler/build-tools/kotlin-build-tools-impl/src/main/kotlin/org/jetbrains/kotlin/buildtools/internal/jvm/operations/JvmCompilationOperationImpl.kt
