+++ compiler/build-tools/kotlin-build-tools-api-forward-tests/build.gradle.kts
@@ -1,4 +1,3 @@
-import org.gradle.kotlin.dsl.invoke
 import org.jetbrains.kotlin.tooling.core.KotlinToolingVersion
 
 plugins {
@@ -172,7 +171,6 @@ val businessLogicTestSuits = setOf(
     "testEscapableCharacters",
     "testInputChangesTracking",
     "testCrossModuleIncrementalChanges",
-    "testFirRunner",
     "testCriToolchain",
     "testCompilerPlugins",
     "testBuildMetrics",
