--- compiler/build-tools/kotlin-build-tools-api-forward-tests/shared/src/compilation/util/scenarioUtil.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright 2010-2025 JetBrains s.r.o. and Kotlin Programming Language contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
- */
-
-package org.jetbrains.kotlin.buildtools.forward.tests.compilation.util
-
-import org.jetbrains.kotlin.buildtools.api.arguments.CommonCompilerArguments
-import org.jetbrains.kotlin.buildtools.api.arguments.ExperimentalCompilerArgument
-import org.jetbrains.kotlin.buildtools.api.jvm.JvmSnapshotBasedIncrementalCompilationConfiguration
-import org.jetbrains.kotlin.buildtools.api.jvm.JvmSnapshotBasedIncrementalCompilationConfiguration.Companion.USE_FIR_RUNNER
-import org.jetbrains.kotlin.buildtools.api.jvm.operations.JvmCompilationOperation
-import org.jetbrains.kotlin.buildtools.forward.tests.compilation.scenario.Scenario
-import org.jetbrains.kotlin.buildtools.forward.tests.compilation.scenario.ScenarioModule
-
-@OptIn(ExperimentalCompilerArgument::class)
-fun Scenario<JvmCompilationOperation.Builder, JvmSnapshotBasedIncrementalCompilationConfiguration.Builder>.moduleWithFir(
-    moduleName: String,
-    dependencies: List<ScenarioModule> = emptyList(),
-    compilationOperationConfig: (JvmCompilationOperation.Builder) -> Unit = {},
-) = module(
-    moduleName = moduleName,
-    dependencies = dependencies,
-    compilationConfigAction = {
-        it.compilerArguments[CommonCompilerArguments.X_USE_FIR_IC] = true
-        compilationOperationConfig(it)
-    },
-    icOptionsConfigAction = {
-        @Suppress("DEPRECATION_ERROR")
-        it[USE_FIR_RUNNER] = true
-    }
-)
