+++ compiler/build-tools/kotlin-build-tools-api-forward-tests/src/testDefaultOptions/kotlin/ModuleNameCompilationTest.kt
@@ -5,10 +5,7 @@
 
 package org.jetbrains.kotlin.buildtools.forward.tests.defaults
 
-import org.jetbrains.kotlin.buildtools.api.arguments.CommonCompilerArguments
-import org.jetbrains.kotlin.buildtools.api.arguments.ExperimentalCompilerArgument
 import org.jetbrains.kotlin.buildtools.api.arguments.JvmCompilerArguments.Companion.MODULE_NAME
-import org.jetbrains.kotlin.buildtools.api.jvm.JvmSnapshotBasedIncrementalCompilationConfiguration.Companion.USE_FIR_RUNNER
 import org.jetbrains.kotlin.buildtools.forward.tests.CompilerExecutionStrategyConfiguration
 import org.jetbrains.kotlin.buildtools.forward.tests.compilation.BaseCompilationTest
 import org.jetbrains.kotlin.buildtools.forward.tests.compilation.assertions.assertCompiledSources
@@ -43,30 +40,6 @@ class ModuleNameCompilationTest : BaseCompilationTest() {
     fun incremental(strategyConfig: CompilerExecutionStrategyConfiguration) {
         jvmScenario(strategyConfig) {
             val module1 = module("basic-multimodule-project/module-1", compilationConfigAction = {
-                it.compilerArguments[MODULE_NAME] = EXPLICIT_NULL_MODULE_NAME_MARKER
-            })
-
-            module1.createPredefinedFile("secret.kt", "new-file")
-            module1.compile {
-                assertCompiledSources("secret.kt")
-                // SecretKt is added, BazKt is removed
-                assertOutputs("META-INF/main.kotlin_module", "SecretKt.class", "Bar.class", "BazKt.class", "FooKt.class")
-                assertModuleNameIsNotSet()
-            }
-        }
-    }
-
-    @OptIn(ExperimentalCompilerArgument::class)
-    @DisplayName("FIR Incremental compilation without specified -module-name")
-    @BtaV2StrategyAgnosticCompilationTest
-    @TestMetadata("basic-multimodule-project/module-1")
-    fun incrementalWithFirRunner(strategyConfig: CompilerExecutionStrategyConfiguration) {
-        jvmScenario(strategyConfig) {
-            val module1 = module("basic-multimodule-project/module-1", icOptionsConfigAction = {
-                @Suppress("DEPRECATION_ERROR")
-                it[USE_FIR_RUNNER] = true
-            }, compilationConfigAction = {
-                it.compilerArguments[CommonCompilerArguments.X_USE_FIR_IC] = true
                 it.compilerArguments[MODULE_NAME] = EXPLICIT_NULL_MODULE_NAME_MARKER
             })
