--- compiler/build-tools/kotlin-build-tools-api-forward-tests/src/testFirRunner/kotlin/LocalClassesFirTest.kt
@@ -1,49 +1,0 @@
-/*
- * Copyright 2010-2026 JetBrains s.r.o. and Kotlin Programming Language contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
- */
-package org.jetbrains.kotlin.buildtools.forward.tests.compilation
-
-import org.jetbrains.kotlin.buildtools.forward.tests.CompilerExecutionStrategyConfiguration
-import org.jetbrains.kotlin.buildtools.forward.tests.compilation.assertions.assertLogContainsPatterns
-import org.jetbrains.kotlin.buildtools.forward.tests.compilation.model.BtaV2StrategyAgnosticCompilationTest
-import org.jetbrains.kotlin.buildtools.forward.tests.compilation.model.LogLevel
-import org.jetbrains.kotlin.buildtools.forward.tests.compilation.scenario.jvmScenario
-import org.jetbrains.kotlin.buildtools.forward.tests.compilation.util.compile
-import org.jetbrains.kotlin.buildtools.forward.tests.compilation.util.moduleWithFir
-import org.jetbrains.kotlin.test.TestMetadata
-import org.junit.jupiter.api.DisplayName
-
-class LocalClassesFirTest : BaseCompilationTest() {
-    @DisplayName("Changing super type of local class should trigger rebuild of the local class")
-    @BtaV2StrategyAgnosticCompilationTest
-    @TestMetadata("ic-scenarios/KT-85074")
-    fun localClassSuperTypeChange(strategyConfig: CompilerExecutionStrategyConfiguration) {
-        jvmScenario(strategyConfig) {
-            val module = moduleWithFir("ic-scenarios/KT-85074")
-
-            module.replaceFileWithVersion("Foo.kt", "addVal")
-
-            module.compile {
-                expectFail()
-                assertLogContainsPatterns(
-                    LogLevel.ERROR,
-                    ".*Class '<anonymous>' is not abstract and does not implement abstract member:\nval bar: String".toRegex()
-                )
-            }
-        }
-    }
-
-    @DisplayName("Changing local class should not trigger rebuild of usages")
-    @BtaV2StrategyAgnosticCompilationTest
-    @TestMetadata("ic-scenarios/local-class-change")
-    fun localClassChange(strategyConfig: CompilerExecutionStrategyConfiguration) {
-        jvmScenario(strategyConfig) {
-            val module = moduleWithFir("ic-scenarios/local-class-change")
-
-            module.replaceFileWithVersion("Foo.kt", "changeLocalClass")
-
-            module.compile(expectedDirtySet = setOf("Foo.kt"))
-        }
-    }
-}
