+++ compiler/build-tools/kotlin-build-tools-api-tests/build.gradle.kts
@@ -222,7 +222,6 @@ val businessLogicTestSuits = setOf(
     "testEscapableCharacters",
     "testInputChangesTracking",
     "testCrossModuleIncrementalChanges",
-    "testFirRunner",
     "testCriToolchain",
     "testCompilerPlugins",
     "testBuildMetrics",
