+++ compiler/build-tools/kotlin-build-tools-api-tests/src/testCrossModuleIncrementalChanges/kotlin/MultiProjectIncrementalChangesTest.kt
@@ -17,11 +17,39 @@ import org.junit.jupiter.api.DisplayName
  * two-module `lib` (depended on by) `app` project. Verifies which sources are recompiled when a declaration in
  * `lib` changes. The Gradle-specific scenarios (`:lib:clean`, dependency/classpath edits, build-dir remapping,
  * missing-IC-state and cache-corruption fallback, Groovy interop, task up-to-date wiring) stay in the integration
- * test. The ABI-method-signature case already lives in [ClassicMultiprojectFirRunnerIncrementalTest].
+ * test.
  */
 class MultiProjectIncrementalChangesTest : BaseCompilationTest() {
     private val aKt = "src/main/kotlin/bar/A.kt"
     private val barDummyKt = "src/main/kotlin/bar/BarDummy.kt"
+
+    @DefaultStrategyAgnosticCompilationTest
+    @DisplayName("Changing a lib method signature recompiles its lib and app dependents, including its callers")
+    @TestMetadata("incrementalMultiproject")
+    fun testAbiChangeInLib_changeMethodSignature(strategyConfig: CompilerExecutionStrategyConfiguration) {
+        jvmScenario(strategyConfig) {
+            val lib = module("incrementalMultiproject/lib")
+            val app = module("incrementalMultiproject/app", dependencies = listOf(lib))
+
+            lib.changeFile(aKt) { it.replace("fun a() {}", "fun a(): Int = 1") }
+
+            lib.compile(
+                expectedDirtySet = setOf(
+                    "src/main/kotlin/bar/A.kt",
+                    "src/main/kotlin/bar/B.kt",
+                    "src/main/kotlin/bar/barUseA.kt",
+                )
+            )
+            app.compile(
+                expectedDirtySet = setOf(
+                    "src/main/kotlin/foo/AA.kt",
+                    "src/main/kotlin/foo/AAA.kt",
+                    "src/main/kotlin/foo/BB.kt",
+                    "src/main/kotlin/foo/fooUseA.kt",
+                )
+            )
+        }
+    }
 
     @DefaultStrategyAgnosticCompilationTest
     @DisplayName("Adding a new method to a lib class recompiles its lib and app dependents")
