--- compiler/build-tools/kotlin-build-tools-api-tests/src/testFirRunner/kotlin/SingleModuleFirRunnerIncrementalTest.kt
@@ -1,101 +1,0 @@
-/*
- * Copyright 2010-2025 JetBrains s.r.o. and Kotlin Programming Language contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
- */
-package org.jetbrains.kotlin.buildtools.tests.compilation
-
-import org.jetbrains.kotlin.buildtools.api.arguments.CommonCompilerArguments.Companion.LANGUAGE_VERSION
-import org.jetbrains.kotlin.buildtools.api.arguments.ExperimentalCompilerArgument
-import org.jetbrains.kotlin.buildtools.api.arguments.enums.KotlinVersion
-import org.jetbrains.kotlin.buildtools.api.jvm.JvmSnapshotBasedIncrementalCompilationConfiguration.Companion.USE_FIR_RUNNER
-import org.jetbrains.kotlin.buildtools.tests.CompilerExecutionStrategyConfiguration
-import org.jetbrains.kotlin.buildtools.tests.compilation.assertions.assertCompiledSources
-import org.jetbrains.kotlin.buildtools.tests.compilation.assertions.assertNoCompiledSources
-import org.jetbrains.kotlin.buildtools.tests.compilation.model.DefaultStrategyAgnosticCompilationTest
-import org.jetbrains.kotlin.buildtools.tests.compilation.scenario.assertAddedOutputs
-import org.jetbrains.kotlin.buildtools.tests.compilation.scenario.assertRemovedOutputs
-import org.jetbrains.kotlin.buildtools.tests.compilation.scenario.jvmScenario
-import org.jetbrains.kotlin.buildtools.tests.compilation.util.moduleWithFir
-import org.jetbrains.kotlin.test.TestMetadata
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.assertThrows
-import java.util.*
-
-
-@DisplayName("Single module IC scenarios for FIR runner")
-class SingleModuleFirRunnerIncrementalTest : BaseCompilationTest() {
-
-    @DefaultStrategyAgnosticCompilationTest
-    @DisplayName("Adding and removing the class")
-    @TestMetadata("basic-multimodule-project/module-1")
-    fun testScenario1(strategyConfig: CompilerExecutionStrategyConfiguration) {
-        jvmScenario(strategyConfig) {
-            val module1 = moduleWithFir("basic-multimodule-project/module-1")
-
-            val randomString = UUID.randomUUID().toString()
-            module1.createFile(
-                "foobar.kt",
-                //language=kt
-                """
-                fun foobar() {
-                    println("$randomString")
-                }
-                """.trimIndent()
-            )
-
-            module1.compile {
-                assertCompiledSources("foobar.kt")
-                assertAddedOutputs("FoobarKt.class") // specify only the difference
-            }
-
-            module1.deleteFile(
-                "foobar.kt",
-            )
-
-            module1.compile {
-                assertNoCompiledSources()
-                assertRemovedOutputs("FoobarKt.class") // specify only the difference
-            }
-        }
-    }
-
-    @DefaultStrategyAgnosticCompilationTest
-    @DisplayName("Throws an exception on using LV 1.9")
-    @TestMetadata("basic-multimodule-project/module-1")
-    fun testScenario2(strategyConfig: CompilerExecutionStrategyConfiguration) {
-        jvmScenario(strategyConfig) {
-            assertThrows<IllegalStateException>(
-                message = "Compilation does not fail on LV 1.9"
-            ) {
-                // Throws on initial compilation
-                moduleWithFir(
-                    moduleName = "basic-multimodule-project/module-1",
-                    compilationOperationConfig = {
-                        it.compilerArguments[LANGUAGE_VERSION] = KotlinVersion.V1_9
-                    },
-                )
-            }
-        }
-    }
-
-    @OptIn(ExperimentalCompilerArgument::class)
-    @DefaultStrategyAgnosticCompilationTest
-    @DisplayName("Throws an exception on missing -Xuse-fir-ic")
-    @TestMetadata("basic-multimodule-project/module-1")
-    fun testScenario3(strategyConfig: CompilerExecutionStrategyConfiguration) {
-        jvmScenario(strategyConfig) {
-            assertThrows<IllegalStateException>(
-                message = "Compilation does not fail on missing -Xuse-fir-ic"
-            ) {
-                // Throws on initial compilation
-                module(
-                    moduleName = "basic-multimodule-project/module-1",
-                    icOptionsConfigAction = {
-                        @Suppress("DEPRECATION_ERROR")
-                        it[USE_FIR_RUNNER] = true
-                    },
-                )
-            }
-        }
-    }
-}
