--- compiler/build-tools/kotlin-build-tools-api-tests/src/testFirRunner/kotlin/TypealiasChangeTest.kt
@@ -1,66 +1,0 @@
-/*
- * Copyright 2010-2025 JetBrains s.r.o. and Kotlin Programming Language contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
- */
-package org.jetbrains.kotlin.buildtools.tests.compilation
-
-import org.jetbrains.kotlin.buildtools.tests.CompilerExecutionStrategyConfiguration
-import org.jetbrains.kotlin.buildtools.tests.compilation.assertions.assertCompiledSources
-import org.jetbrains.kotlin.buildtools.tests.compilation.model.DefaultStrategyAgnosticCompilationTest
-import org.jetbrains.kotlin.buildtools.tests.compilation.scenario.jvmScenario
-import org.jetbrains.kotlin.buildtools.tests.compilation.util.moduleWithFir
-import org.jetbrains.kotlin.test.TestMetadata
-import org.junit.jupiter.api.Disabled
-import org.junit.jupiter.api.DisplayName
-
-
-@DisplayName("Single module IC scenarios with typealiases")
-class TypealiasChangeTest : BaseCompilationTest() {
-
-    @Disabled("Broken, KT-28233")
-    @DefaultStrategyAgnosticCompilationTest
-    @DisplayName("Potential first-round errors: typealias change")
-    @TestMetadata("empty")
-    fun testTypealiasChange(strategyConfig: CompilerExecutionStrategyConfiguration) {
-        jvmScenario(strategyConfig) {
-            val module = moduleWithFir("empty")
-
-            module.createFile(
-                "Foo.kt",
-                """
-                    interface Foo {
-                        val values: Type
-                    }
-                """.trimIndent()
-            )
-
-            module.createFile(
-                "FooImpl.kt",
-                """
-                    class FooImpl : Foo {
-                        override val values: Type
-                            get() = "0"
-                    }
-                """.trimIndent()
-            )
-
-            module.createFile(
-                "types.kt",
-                """
-                    typealias Type = String
-                """.trimIndent()
-            )
-
-            module.compile {
-                assertCompiledSources("Foo.kt", "FooImpl.kt", "types.kt")
-            }
-
-            module.changeFile("types.kt") { contents -> contents.replace("String", "Int") }
-            module.changeFile("FooImpl.kt") { contents -> contents.replace("\"0\"", "0") }
-
-            module.compile {
-                assertCompiledSources("Foo.kt", "FooImpl.kt", "types.kt")
-            }
-        }
-    }
-}
