+++ repo/testLifecycleTask.dump.txt
@@ -27,8 +27,6 @@ This dump is generated by TestLifecycleTaskTest.kt
   - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testEscapableCharacters2.4.20-Beta1 [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testExample2.4.0 [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testExample2.4.20-Beta1 [BuildToolsApi]
-  - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testFirRunner2.4.0 [BuildToolsApi]
-  - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testFirRunner2.4.20-Beta1 [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testForwardCompatibility2.4.0 [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testForwardCompatibility2.4.20-Beta1 [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testInputChangesTracking2.4.0 [BuildToolsApi]
@@ -57,7 +55,6 @@ This dump is generated by TestLifecycleTaskTest.kt
   - :compiler:build-tools:kotlin-build-tools-api-tests:testDaemonOptions [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-tests:testDefaultOptions [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-tests:testEscapableCharacters [BuildToolsApi]
-  - :compiler:build-tools:kotlin-build-tools-api-tests:testFirRunner [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-tests:testInputChangesTracking [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-tests:testInternalInputsTracker [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-tests:testKotlinLogger [BuildToolsApi]
@@ -202,8 +199,6 @@ This dump is generated by TestLifecycleTaskTest.kt
   - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testEscapableCharacters2.4.20-Beta1 [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testExample2.4.0 [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testExample2.4.20-Beta1 [BuildToolsApi]
-  - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testFirRunner2.4.0 [BuildToolsApi]
-  - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testFirRunner2.4.20-Beta1 [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testForwardCompatibility2.4.0 [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testForwardCompatibility2.4.20-Beta1 [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-forward-tests:testInputChangesTracking2.4.0 [BuildToolsApi]
@@ -232,7 +227,6 @@ This dump is generated by TestLifecycleTaskTest.kt
   - :compiler:build-tools:kotlin-build-tools-api-tests:testDaemonOptions [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-tests:testDefaultOptions [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-tests:testEscapableCharacters [BuildToolsApi]
-  - :compiler:build-tools:kotlin-build-tools-api-tests:testFirRunner [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-tests:testInputChangesTracking [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-tests:testInternalInputsTracker [BuildToolsApi]
   - :compiler:build-tools:kotlin-build-tools-api-tests:testKotlinLogger [BuildToolsApi]
